[FLINK-14081][table-planner-blink] fix PartitionPruner in blink planner and TimeSerializer in flink python

+++ flink-python/src/main/java/org/apache/flink/table/runtime/typeutils/PythonTypeUtils.java
@@ -264,7 +264,7 @@ public final class PythonTypeUtils {
 
 		@Override
 		public TypeSerializer visit(TimeType timeType) {
-			return IntSerializer.INSTANCE;
+			return LongSerializer.INSTANCE;
 		}
 
 		@Override
